+++ iotile_cloud/utils/mock_cloud.py
@@ -76,6 +76,9 @@ class MockIOTileCloud(object):
 
         # APIs for getting raw data
         self._add_api(r"/api/v1/stream/(s--[0-9\-a-f]+)/data/", self.get_stream_data)
+        # Function will get stream from '?filter='
+        self._add_api(r"/api/v1/data/", self.get_stream_data)
+        self._add_api(r"/api/v1/df/", self.get_stream_df)
         self._add_api(r"/api/v1/event/([0-9]+)/data/", self.get_raw_event)
 
         # APIs for querying single models
@@ -363,8 +366,17 @@ class MockIOTileCloud(object):
 
         return results
 
-    def get_stream_data(self, request, stream):
-        if stream not in self.streams:
+    def get_stream_data(self, request, stream=None):
+
+        if not stream:
+            # Stream can only be none if not used by /stream/<stream>/
+            # which means this is called from either /df/?filter= or /data/?filter=
+            if 'filter' in request.args:
+                filter = request.args['filter']
+                if filter.split('--')[0] == 's':
+                    stream = filter + '.raw'
+
+        if stream.split('.')[0] not in self.streams:
             raise ErrorCode(404)
 
         results = []
@@ -380,6 +392,21 @@ class MockIOTileCloud(object):
                 results = self._format_stream_data(self.streams[stream], csv_stream_path)
 
         return self._paginate(results, request, 1000)
+
+    def get_stream_df(self, request, stream=None):
+
+        results = []
+
+        data = self.get_stream_data(request, stream)
+
+        for row in data['results']:
+            results.append({
+                'row': row['timestamp'],
+                'value': row['value'],
+                'stream_slug': row['stream']
+            })
+
+        return results
 
     def _format_stream_data(self, stream, csvpath):
         results = []